migrate NetworkObject to macro
